// src/client/types.gen.ts
export type PublicPageMoveOrCopyData = {
	body?: {
		/**
		 * ID of target parent page (optional)
		 */
		parentId?: number | null
		/**
		 * Target title (optional)
		 */
		title?: string | null
		/**
		 * Index in subpage order (optional, default 0)
		 */
		index?: number | null
		/**
		 * Copy the page instead of move (optional, default false)
		 */
		copy?: boolean
	}
	headers: {
		/**
		 * Required to be true for the API request to pass
		 */
		'OCS-APIRequest': boolean
	}
	path: {
		apiVersion: '1.0'
		token: string
		/**
		 * ID of the page
		 */
		id: number
	}
	query?: never
	url: '/ocs/v2.php/apps/collectives/api/v{apiVersion}/p/collectives/{token}/pages/{id}'
}
// ...
export type PublicPageCreateData = {
	body: {
		/**
		 * Title of the page
		 */
		title: string
		/**
		 * ID of the template page to use
		 */
		templateId?: number | null
	}
	headers: {
		/**
		 * Required to be true for the API request to pass
		 */
		'OCS-APIRequest': boolean
	}
	path: {
		apiVersion: '1.0'
		token: string
		/**
		 * ID of the parent page
		 */
		parentId: number
	}
	query?: never
	url: '/ocs/v2.php/apps/collectives/api/v{apiVersion}/p/collectives/{token}/pages/{parentId}'
}